compare azure blobs
