+++ label_studio/io_storages/azure_blob/api.py
@@ -2,7 +2,7 @@
 """
 from django.utils.decorators import method_decorator
 from drf_yasg import openapi
-from drf_yasg.utils import no_body, swagger_auto_schema
+from drf_yasg.utils import swagger_auto_schema
 from io_storages.api import (
     ExportStorageDetailAPI,
     ExportStorageFormLayoutAPI,
@@ -15,24 +15,20 @@ from io_storages.api import (
     ImportStorageSyncAPI,
     ImportStorageValidateAPI,
 )
-from io_storages.azure_blob.models import AzureBlobExportStorage, AzureBlobImportStorage
-from io_storages.azure_blob.serializers import AzureBlobExportStorageSerializer, AzureBlobImportStorageSerializer
-
-from .openapi_schema import (
-    _azure_blob_export_storage_schema,
-    _azure_blob_export_storage_schema_with_id,
-    _azure_blob_import_storage_schema,
-    _azure_blob_import_storage_schema_with_id,
+from io_storages.azure_serviceprincipal.models import (
+    AzureServicePrincipalExportStorage,
+    AzureServicePrincipalImportStorage,
+)
+from io_storages.azure_serviceprincipal.serializers import (
+    AzureServicePrincipalExportStorageSerializer,
+    AzureServicePrincipalImportStorageSerializer,
 )
 
 
 @method_decorator(
     name='get',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['import_storage', 'azure'],
-        x_fern_sdk_method_name='list',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Get all import storage',
         operation_description='Get list of all Azure import storage connections.',
         manual_parameters=[
@@ -43,150 +39,102 @@ from .openapi_schema import (
                 description='Project ID',
             ),
         ],
-        request_body=no_body,
     ),
 )
 @method_decorator(
     name='post',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['import_storage', 'azure'],
-        x_fern_sdk_method_name='create',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Create new storage',
-        operation_description='Create new Azure import storage',
-        request_body=_azure_blob_import_storage_schema,
+        operation_description='Get new Azure import storage',
     ),
 )
-class AzureBlobImportStorageListAPI(ImportStorageListAPI):
-    queryset = AzureBlobImportStorage.objects.all()
-    serializer_class = AzureBlobImportStorageSerializer
+class AzureServicePrincipalImportStorageListAPI(ImportStorageListAPI):
+    queryset = AzureServicePrincipalImportStorage.objects.all()
+    serializer_class = AzureServicePrincipalImportStorageSerializer
 
 
 @method_decorator(
     name='get',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['import_storage', 'azure'],
-        x_fern_sdk_method_name='get',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Get import storage',
         operation_description='Get a specific Azure import storage connection.',
-        request_body=no_body,
     ),
 )
 @method_decorator(
     name='patch',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['import_storage', 'azure'],
-        x_fern_sdk_method_name='update',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Update import storage',
         operation_description='Update a specific Azure import storage connection.',
-        request_body=_azure_blob_import_storage_schema,
     ),
 )
 @method_decorator(
     name='delete',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['import_storage', 'azure'],
-        x_fern_sdk_method_name='delete',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Delete import storage',
         operation_description='Delete a specific Azure import storage connection.',
-        request_body=no_body,
     ),
 )
-class AzureBlobImportStorageDetailAPI(ImportStorageDetailAPI):
-    queryset = AzureBlobImportStorage.objects.all()
-    serializer_class = AzureBlobImportStorageSerializer
+class AzureServicePrincipalImportStorageDetailAPI(ImportStorageDetailAPI):
+    queryset = AzureServicePrincipalImportStorage.objects.all()
+    serializer_class = AzureServicePrincipalImportStorageSerializer
 
 
 @method_decorator(
     name='post',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['import_storage', 'azure'],
-        x_fern_sdk_method_name='sync',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Sync import storage',
         operation_description='Sync tasks from an Azure import storage connection.',
-        manual_parameters=[
-            openapi.Parameter(
-                name='id',
-                type=openapi.TYPE_INTEGER,
-                in_=openapi.IN_PATH,
-                description='Storage ID',
-            ),
-        ],
-        request_body=no_body,
     ),
 )
-class AzureBlobImportStorageSyncAPI(ImportStorageSyncAPI):
-    serializer_class = AzureBlobImportStorageSerializer
+class AzureServicePrincipalImportStorageSyncAPI(ImportStorageSyncAPI):
+    serializer_class = AzureServicePrincipalImportStorageSerializer
 
 
 @method_decorator(
     name='post',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['export_storage', 'azure'],
-        x_fern_sdk_method_name='sync',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Sync export storage',
         operation_description='Sync tasks from an Azure export storage connection.',
-        request_body=no_body,
     ),
 )
-class AzureBlobExportStorageSyncAPI(ExportStorageSyncAPI):
-    serializer_class = AzureBlobExportStorageSerializer
+class AzureServicePrincipalExportStorageSyncAPI(ExportStorageSyncAPI):
+    serializer_class = AzureServicePrincipalExportStorageSerializer
 
 
 @method_decorator(
     name='post',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['import_storage', 'azure'],
-        x_fern_sdk_method_name='validate',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Validate import storage',
         operation_description='Validate a specific Azure import storage connection.',
-        request_body=_azure_blob_import_storage_schema_with_id,
-        # expecting empty response
-        responses={200: openapi.Response(description='OK')},
     ),
 )
-class AzureBlobImportStorageValidateAPI(ImportStorageValidateAPI):
-    serializer_class = AzureBlobImportStorageSerializer
+class AzureServicePrincipalImportStorageValidateAPI(ImportStorageValidateAPI):
+    serializer_class = AzureServicePrincipalImportStorageSerializer
 
 
 @method_decorator(
     name='post',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['export_storage', 'azure'],
-        x_fern_sdk_method_name='validate',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Validate export storage',
         operation_description='Validate a specific Azure export storage connection.',
-        request_body=_azure_blob_export_storage_schema_with_id,
-        # expecting empty response
-        responses={200: openapi.Response(description='OK')},
     ),
 )
-class AzureBlobExportStorageValidateAPI(ExportStorageValidateAPI):
-    serializer_class = AzureBlobExportStorageSerializer
+class AzureServicePrincipalExportStorageValidateAPI(ExportStorageValidateAPI):
+    serializer_class = AzureServicePrincipalExportStorageSerializer
 
 
 @method_decorator(
     name='get',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['export_storage', 'azure'],
-        x_fern_sdk_method_name='list',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Get all export storage',
         operation_description='Get a list of all Azure export storage connections.',
         manual_parameters=[
@@ -197,70 +145,53 @@ class AzureBlobExportStorageValidateAPI(ExportStorageValidateAPI):
                 description='Project ID',
             ),
         ],
-        request_body=no_body,
     ),
 )
 @method_decorator(
     name='post',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['export_storage', 'azure'],
-        x_fern_sdk_method_name='create',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Create export storage',
         operation_description='Create a new Azure export storage connection to store annotations.',
-        request_body=_azure_blob_export_storage_schema,
     ),
 )
-class AzureBlobExportStorageListAPI(ExportStorageListAPI):
-    queryset = AzureBlobExportStorage.objects.all()
-    serializer_class = AzureBlobExportStorageSerializer
+class AzureServicePrincipalExportStorageListAPI(ExportStorageListAPI):
+    queryset = AzureServicePrincipalExportStorage.objects.all()
+    serializer_class = AzureServicePrincipalExportStorageSerializer
 
 
 @method_decorator(
     name='get',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['export_storage', 'azure'],
-        x_fern_sdk_method_name='get',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Get export storage',
         operation_description='Get a specific Azure export storage connection.',
-        request_body=no_body,
     ),
 )
 @method_decorator(
     name='patch',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['export_storage', 'azure'],
-        x_fern_sdk_method_name='update',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Update export storage',
         operation_description='Update a specific Azure export storage connection.',
-        request_body=_azure_blob_export_storage_schema,
     ),
 )
 @method_decorator(
     name='delete',
     decorator=swagger_auto_schema(
-        tags=['Storage: Azure'],
-        x_fern_sdk_group_name=['export_storage', 'azure'],
-        x_fern_sdk_method_name='delete',
-        x_fern_audiences=['public'],
+        tags=['Storage: Azure SPI'],
         operation_summary='Delete export storage',
         operation_description='Delete a specific Azure export storage connection.',
-        request_body=no_body,
     ),
 )
-class AzureBlobExportStorageDetailAPI(ExportStorageDetailAPI):
-    queryset = AzureBlobExportStorage.objects.all()
-    serializer_class = AzureBlobExportStorageSerializer
+class AzureServicePrincipalExportStorageDetailAPI(ExportStorageDetailAPI):
+    queryset = AzureServicePrincipalExportStorage.objects.all()
+    serializer_class = AzureServicePrincipalExportStorageSerializer
 
 
-class AzureBlobImportStorageFormLayoutAPI(ImportStorageFormLayoutAPI):
+class AzureServicePrincipalImportStorageFormLayoutAPI(ImportStorageFormLayoutAPI):
     pass
 
 
-class AzureBlobExportStorageFormLayoutAPI(ExportStorageFormLayoutAPI):
+class AzureServicePrincipalExportStorageFormLayoutAPI(ExportStorageFormLayoutAPI):
     pass
