+++ label_studio/io_storages/azure_blob/utils.py
@@ -1,12 +1,31 @@
 import fnmatch
 import logging
+import os
 import re
 
 from azure.storage.blob import BlobServiceClient
 from core.utils.params import get_env
+from cryptography.fernet import Fernet
 from django.conf import settings
 
+key = os.getenv('ENCRYPT_KEY', 'CEcJNMyffg-wHysgcW0xaYleNQt9o2LExxsEgB7GkD8=')
+secret_encrypter = Fernet(key.encode())
+
 logger = logging.getLogger(__name__)
+
+
+def get_secured(value):
+    if value is not None:
+        # we decrypt the value
+        value = secret_encrypter.decrypt(str(value).encode()).decode()
+    return value
+
+
+def set_secured(value):
+    if value is not None:
+        # we decrypt the value
+        value = secret_encrypter.encrypt(str(value).encode()).decode()
+    return value
 
 
 class AZURE(object):
