--- label_studio/io_storages/azure_serviceprincipal/__init__.py
@@ -1,2 +1,0 @@
-"""This file and its contents are licensed under the Apache License 2.0. Please see the included NOTICE for copyright information and LICENSE for a copy of the license.
-"""
